web API issues resolved for Python 3.6
The web API sort of works fine on Python 3.6 - the first issue was that previously the output to the web browser wasn't encoded to utf-8, which led to the code not working via this interface. The second issue at the moment were numerous AttributeError messages in the terminal when using the web API - solved via a lousy except-pass statement.